# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bronze Backend API",
    "description": "A comprehensive Go backend with MinIO integration, featuring file processing, parallel job management, and archive decompression capabilities",
    "version": "1.0.0",
    "contact": {
      "name": "Bronze Backend Team",
      "email": "support@bronze-backend.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8060",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/jobs": {
      "post": {
        "summary": "Create job",
        "description": "Create a new processing job",
        "tags": ["Jobs"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Job created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List jobs",
        "description": "List all jobs with optional status filtering",
        "tags": ["Jobs"],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter jobs by status (pending, processing, completed, failed, cancelled)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["pending", "processing", "completed", "failed", "cancelled"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Jobs retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobsListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/jobs/{id}": {
      "get": {
        "summary": "Get job",
        "description": "Get specific job details",
        "tags": ["Jobs"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Job ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "404": {
            "description": "Job not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Cancel job",
        "description": "Cancel a specific job",
        "tags": ["Jobs"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Job ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job cancelled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "404": {
            "description": "Job not found or cannot be cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Error description"
          },
          "error": {
            "type": "string",
            "example": "Detailed error message"
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          }
        }
      },
...
      "Job": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "type": {
            "type": "string",
            "example": "decompress"
          },
          "priority": {
            "type": "string",
            "enum": ["high", "medium", "low"],
            "example": "high"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "processing", "completed", "failed", "cancelled"],
            "example": "pending"
          },
          "file_path": {
            "type": "string",
            "example": "/tmp/example.zip"
          },
          "bucket": {
            "type": "string",
            "example": "files"
          },
          "object_name": {
            "type": "string",
            "example": "example.zip"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2023-01-01T00:00:00Z"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "error": {
            "type": "string",
            "nullable": true
          },
          "result": {
            "type": "object",
            "nullable": true
          },
          "progress": {
            "type": "number",
            "format": "float",
            "minimum": 0,
            "maximum": 100,
            "example": 50
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "CreateJobRequest": {
        "type": "object",
        "required": ["type", "file_path", "bucket", "object_name"],
        "properties": {
          "type": {
            "type": "string",
            "description": "Job type",
            "example": "decompress"
          },
          "file_path": {
            "type": "string",
            "description": "Path to the file to process",
            "example": "/tmp/example.zip"
          },
          "bucket": {
            "type": "string",
            "description": "MinIO bucket name",
            "example": "files"
          },
          "object_name": {
            "type": "string",
            "description": "Object name in MinIO",
            "example": "example.zip"
          },
          "priority": {
            "type": "string",
            "enum": ["high", "medium", "low"],
            "description": "Job priority",
            "default": "medium",
            "example": "high"
          }
        }
      },
      "JobResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessResponse" },
          {
            "type": "object",
            "properties": {
              "job": {
                "$ref": "#/components/schemas/Job"
              }
            }
          }
        ]
      },
      "JobsListResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessResponse" },
          {
            "type": "object",
            "properties": {
              "jobs": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Job"
                }
              },
              "count": {
                "type": "integer",
                "example": 5
              }
            }
          }
        ]
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}